requirements.txt
CherryPy==17.3.0
envparse
pygithub[integrations]
requests==2.19.1
